extend session ID by exec ID and shrink instance ID
The exec ID is added.
The instance ID is shrunk from 44 bits to 32.
Add tests.

+++ mpc/src/ffi/c_bindings/honey_badger_mpc_client/mod.rs
@@ -75,7 +75,7 @@ pub extern "C" fn new_honey_badger_mpc_client(
     id: usize,
     n: usize,
     t: usize,
-    instance_id: u64,
+    instance_id: u32,
     inputs: U256Slice,
     input_len: usize,
     field_kind: FieldKind,
